src/third_party/blink/web_tests/css3/flexbox/flex-item-stretch-image.html

<!DOCTYPE html>
<title>CSS Flexbox: image stretch</title>
<meta name="assert" content="This test ensures that flexbox stretches a given image
to fit the size of flexitem according to flex values.">
<link rel="issue" href="https://bugs.chromium.org/p/chromium/issues/detail?id=721123">
<link rel="bookmark" href="http://wpt.live/css/css-flexbox/flexitem-stretch-image.html">
<html>
<style>
.flexbox {
    width: 600px;
    display: flex;
    background-color: #aaa;
    position: relative;
    min-height: 10px;
}
.flexbox > * {
    margin: 0;
    border: 0;
    padding: 0;
    min-width: 0;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<!-- Blink's behavior here is incorrect, see https://crbug.com/721123.
     When that bug is fixed, please delete this test, as this is correctly
     tested by external/wpt/css/css-flexbox/flexitem-stretch-image.html -->

<div class="flexbox">
  <img data-expected-display="block" data-expected-width="345" style="flex: 1 0 auto;" src="../../images/resources/blue-100.png">
  <img data-expected-display="block" data-expected-width="255" data-expected-height="100" style="flex: 1 0 auto;" src="../../images/resources/green-10.png">
</div>

</body>
